Fix persist Marshaling

+++ lib/supermodel/persist.rb
@@ -14,16 +14,22 @@ module SuperModel
     
     def load
       return unless path
+      return unless File.exist?(path)
+      records = []
       File.open(path, "rb") {|file|
-        Marshal.load(file)
+        records = Marshal.load(file)
       }
+      records.flatten.each {|r| r.class.records << r }
+      true
     end
   
     def dump
       return unless path
+      records = klasses.map {|k| k.records }
       File.open(path, "wb+") {|file|
-        Marshal.dump(klasses, file)
+        Marshal.dump(records, file)
       }
+      true
     end
     
     extend self
@@ -31,6 +37,14 @@ module SuperModel
     module Model
       def self.included(base)
         Persist.klasses << base
+      end
+      
+      def marshal_dump
+        @attributes
+      end
+
+      def marshal_load(attributes)
+        @attributes = attributes
       end
     end
   end
